Remove ROS1 pcl_recorder version
